requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
grpcio>=1.54.0 